etc: 탭 컴퍼넌트 둘로 나눔에 따른 조정
