+++ src/components/index.ts
@@ -1,4 +1,3 @@
-export { default as TabContents } from "./TabContents";
 export { default as Layout } from "./Layout";
 export { default as Footer } from "./Footer";
 export { default as DataDummy } from "./DataDummy";
@@ -8,3 +7,5 @@ export { default as Result } from "./Result";
 export { default as Search } from "./Search";
 export { default as Tags } from "./Tags";
 export { default as MovieSlide } from "./MovieSlide";
+export { default as TabTitle } from "./TabTitle";
+export { default as TabContent } from "./TabContent";
