+++ src/pages/index.ts
@@ -1,4 +1,4 @@
-export { default as Detail } from "./Detail";
 export { default as Home } from "./Home";
 export { default as TestPage } from "./TestPage";
 export { default as SearchPage } from "./SearchPage";
+export { default as Detail } from "./Detail";
